requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-environ==0.7.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
facebook-sdk==3.1.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.8.1
Markdown==3.3.4
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
twilio==7.0.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.5.0
